passed all tests.

+++ tajo-common/src/main/java/org/apache/tajo/datum/BooleanDatum.java
@@ -29,9 +29,9 @@ public class BooleanDatum extends Datum {
   public static final BooleanDatum TRUE = new BooleanDatum(true);
   public static final BooleanDatum FALSE = new BooleanDatum(false);
 
-  public static final int UNKNOWN_INT = 0;
-  public static final int TRUE_INT = 1;
-  public static final int FALSE_INT = 2;
+  public static final byte UNKNOWN_INT = 0;
+  public static final byte TRUE_INT = 1;
+  public static final byte FALSE_INT = 2;
   public static final byte [] TRUE_BYTES = new byte[] {TRUE_INT};
   public static final byte [] FALSE_BYTES = new byte[] {FALSE_INT};
